Adds -i flag for indicator to batchgit
This is a simple implementation so that the -i flag will start the
indicator. There is an issue with how this is done.

If batchgit is not sourced directly from the install location, then this
script will fail currently fail to find the indicator. This is because
it is found using a relative path to batchgit.